+++ backend/libraries/types/src/notifications.rs
@@ -29,7 +29,7 @@ pub struct UserNotification {
     pub notification_bytes: ByteBuf,
 
     // Values relevant for the FCM notifications
-    #[serde(skip_deserializing, rename = "f")]
+    #[serde(rename = "f")]
     pub fcm_data: Option<FcmData>,
 }
 
@@ -153,7 +153,7 @@ pub struct UserNotificationEnvelope {
     pub notification_bytes: ByteBuf,
     #[serde(rename = "t")]
     pub timestamp: TimestampMillis,
-    #[serde(skip_deserializing, rename = "f")]
+    #[serde(default, rename = "f")]
     pub fcm_data: Option<FcmData>,
 }