+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Deprecate `winners` field on prize messages ([#8302](https://github.com/open-chat-labs/open-chat/pull/8302))
 
+### Changed
+- Re-enabled fcm_data ([8298](https://github.com/open-chat-labs/open-chat/pull/8298))
+
+
 ### Removed
 
 - Removed unused fields from BotChatEvent (must go ahead of groups/communities) ([#8291](https://github.com/open-chat-labs/open-chat/pull/8291))